Removed 1 unnecessary stubbing in ClauseActionFactoryTest.java
